chore(release-published.yml): add workflow step to upload xsds to S3 bucket
The commit adds a new step to the release-published.yml workflow file. This step is responsible for uploading xsd files to an S3 bucket. The step first downloads the liquibase mongodb xsd repository using the actions/checkout action. Then, it sets the necessary environment variables for AWS access. Next, it uses the aws s3 sync command to sync the xsd files from the local repository to the specified S3 bucket. Finally, it performs additional operations on the index.htm file, such as adding links for all xsd files (except liquibase-mongodb-latest.xsd and index.htm), and syncs the modified index.htm file with the S3 bucket.