feat(get_draft_release.sh): add script to retrieve draft release information from GitHub API
The script `get_draft_release.sh` is added to the project. This script retrieves information about the draft release from the GitHub API. It takes a single argument `KEY`, which can be either "TAG" or "UPLOAD_URL".

If the `GITHUB_TOKEN` environment variable is not set, the script will display an error message and exit. Similarly, if the `GITHUB_REPOSITORY` environment variable is not set, the script will display an error message and exit.

The script makes a GET request to the GitHub API using the provided `GITHUB_TOKEN` and `GITHUB_REPOSITORY`. It retrieves all releases and filters for the draft release. If a draft release is found, the script performs the following actions based on the value of `KEY`:

- If `KEY` is "TAG", the script extracts the HTML URL of the draft release and outputs the tag name.
- If `KEY` is "UPLOAD_URL", the script extracts the upload URL of the draft release and outputs it.

If no draft release is found, the script displays an error message and exits.

Note: There is no newline at the end of the file.