feat(get_draft_release.sh): add script to retrieve draft release information from GitHub API
The script `get_draft_release.sh` is added to the project. It retrieves information about the draft release from the GitHub API. The script takes a single argument `KEY`, which can be either `TAG` or `UPLOAD_URL`.

If the `GITHUB_TOKEN` environment variable is not set, the script will display an error message and exit. Similarly, if the `GITHUB_REPOSITORY` environment variable is not set, the script will display an error message and exit.

The script makes a GET request to the GitHub API using the provided token and repository information. It retrieves all releases and filters for the draft release. If a draft release is found, the script performs the following actions based on the value of `KEY`:

- If `KEY` is `TAG`, the script extracts the HTML URL of the draft release and extracts the tag name from it. The tag name is then printed to the console.
- If `KEY` is `UPLOAD_URL`, the script extracts the upload URL of the draft release and removes the query parameters from it. The modified upload URL is then printed to the console.

Note: The script does not have a newline character at the end.