chore(get_draft_release.sh): remove unused script for retrieving draft release information
The script `get_draft_release.sh` was removed as it was no longer being used. This script was used to retrieve information about draft releases from the GitHub API using the provided GitHub token and repository. However, it was determined that this functionality was no longer necessary and the script was deleted.